Change Homer template to get license key from custom field
To overcome the 200 character limit on the description field we're
moving the software license keys for network devices to a custom
field of the existing inventory items were storing them in.

This changes the Jinja template to get the key from the custom field
rather than the description.

I've populated the custom-fields on all devices already, once merged
I'll null the description field so the custom field is the only
place the key exists.

Bug: T334180
Change-Id: Ic739294886bce14e4b277797f5b5ca36995258f8