Avoid creating EVPN import policy with default accept if no Vlans
As things stand, if no Vlans exist on a switch, homer templates
produce a 'l2vni' term in the Evpn_rt_import policy with no
communities matched.  This is invalid when applied under
vrf-import in switch-options.

This patch will only include the term if vlans are defined for the
switch and thus communites included in it.

Also adding similar logic for L3 VRF route-targets.  Can't really
imagine a scenario where we've a L2-only EVPN device, but it makes
sense to apply the same logic I think.

Bug: T322937
Change-Id: Ife60f5ad6bf0a8659034583fe72eb9917c9a3529