Don't set routed sub-int parents as L2 trunks on switches
Current config sets all ports with 'mode' set as L2 switch port.

After the merge of changes in netbox to set sub-interface parents
some switch ints (cloudsw and lsw CR uplinks) are now set to
trunk reflecting the presence of routed sub-interfaces, so on
these we don't want to set the parent up as a L2 port.

Bug: T296832
Change-Id: I49e01b4cbec002a8157c9dc5f6c7eb441f589f2e