Add trust-option-82 to dhcp relay conf for core routers
Had a bit of a time getting the new DHCP relay config to work on
the core routers.  Eventually turned out this command was required
on them or they would drop the packets with option 82 in them added
by our L2 switch.

So config is now "if l3switch" - add the option 82 info - if not then
trust the option 82 info already present from L2 switch.

Bug: T320508
Change-Id: Id41ab55e6af5a7bd5c2001fb88a137051298db7b