Disable IPv6 RA generation on spine layer switches
This patch disables generation of IPv6 RA config on switches
with name stating 'ssw', matching our spine layer devices.

As we have routed top-of-rack switches we should never need to
have Spines generate RAs or act as gateway for servers, but
in some cases we do need them to have an interface on vlans.

Bug: T337057
Change-Id: I9ec9ab25bfc0ed1cab5e8848ee598dd2a5e1673d