Add class-of-service parent interface shaper for sub-rated services
This was part of the wider class-of-service configs I'd prepped for
the QoS stuff, however it's a separate part of the config so it
can be introduced on it's own.

Given the seemingly poor performance of the one circuit that needs
this I think it may be best to introduce now and see if it might
help with performance.  In theory the shaper will improve TCP
performance on the service, compared to letting the CR burst
beyond the carrier commit rate and hitting their policer.

Only produces a diff on two routers:

Changes for 1 devices: ['cr1-codfw.wikimedia.org']

[edit]
+  class-of-service {
+      interfaces {
+          xe-1/0/1:2 {
+              shaping-rate 3920000000;
+          }
+      }
+  }

---------------
Changes for 1 devices: ['cr3-eqsin.wikimedia.org']

[edit]
+  class-of-service {
+      interfaces {
+          xe-0/1/0 {
+              shaping-rate 3920000000;
+          }
+      }
+  }

Bug: T337220
Bug: T328313
Change-Id: Ie39a05b04e32f07fe3f5b50faa8a943765c753d7