DHCP: allow transit DHCP from 3rd party relays
The modern DHCP implementation on Juniper devices forwards ALL DHCP
packets to the control plane for further processing by the DHCP deamon
as soon as this deamon is enabled.
Which means even unicast "just cruising" packets from management routers
to install servers gets intercepted.
The fix is to:
1/ permit those packets to enter the control plane (because we have no
choice anyway, we can't tell it not to intercept them)
2/ permit those packets at the dhcp deamon level (from non-configured
interface)

For 1/ it's not possible to see what those loopback packets look like,
and it seems like Junos is re-writing them as filtering on destination
IP (matching the install server or 10/8) doesn't work. Same for
filtering on destination port 67 or 68, both need to be allowed (67-68).

Bug: T337345
Change-Id: I5f340deffb3021449998fb93ca9dd1f15635bf77