remove old gerrit service IP from static definitions
Cleanup after recent migration of gerrit from old
to new hardware which involved a change of the service
IP behind gerrit.wikimedia.org.

208.80.154.137 used to be gerrit.wikimedia.org but is
now gerrit-old.wikimedia.org.

208.80.154.151 is now gerrit.wikimedia.org.

The former prod service IP points to gerrit1001
as https://gerrit-old.wikimedia.org/r/ and the gerrit
service is shutdown, so it shows only the maintenance page.

This is the step we had forgotten during the migration recently
which caused broken CI from cloud for a little time after switching.

Sukhbir had helped us to fix it, which is when we added the new IP
additionally.

So what this really means is "cloud can't talk to gerrit1001 anymore".

Bug: T336427
Change-Id: Ic86af26e867b5b0e137b591f758ed52313967add