Add rule to allow TFTP to install server to support Juniper ZTP
As discussed in chat with Riccardo and Papaul due to timelines and
leave it makes sense to allow TFTP to the install server temporarily
so the new switches can download the ztp configuration script.

Plan is still to make this available over HTTP from the apt server,
at which point we can remove this again.  But as that might not
be ready for a week or two, by which stage papaul will be out, it
was seen as best to allow this for now.

Bug: T336485
Change-Id: I39873dfd62dfbfe7c60e756583667645bebda239