Allow connections from management networks to apt servers on 8080
This change allows devices on our management network make connections
to our APT servers on TCP 8080, which will increase the security in
terms of how we serve Juniper images and ZTP scripts.

I did consider updating the 'WEB' port declaration to also include
8080 in addition to 80 and 443, however this definition is used in
the labs and cloud inbound policies.  Probably not a major issue if
it were included but to be safe I think best we don't open anything
else from those sources.

Bug: T337028
Change-Id: Ibd669469c6fbb1c725fb6d488443c642e13dd6a3