Change DHCP relay function on management routers to 'forward-only'
As described in the task the defaut DHCP snooping that occurs on
our management routers when relaying DHCP packets is causing a
problem for Juniper ZTP on newer versions, as the MAC address
on the switch side changes during the process, conflicting
with the DHCP binding the MR adds.

Changing the relay operation to 'forward-only', which simply relays
DHCP messages to the configured server over unicast IPv4, resolves
the issue.  This is an option on the management side as we don't
need to insert any additional info into DHCP requests like we do
on production networks during reimage.

Bug: T345273
Change-Id: I8dfc3526e0b7e0ebf3e8ef77dd48741b585d7ec5