cr-labs: Remove port filter on Prometheus term
Right now there are 5+ Prometheus jobs that are failing to reach cloud*
nodes because they're not listed on the firewall term here. Between the
high number of ports, and the plans in [1] which would increase that
number even futher, it feels simpler to just remove the port filter
from this firewall term rather than continuing to list all the ports.

[0]: https://w.wiki/7Psu
[1]: https://phabricator.wikimedia.org/T336854

Change-Id: If6cdcbe2fe6df0dfef2fb85bb0d0c8adda3d22eb