Adjust hashing algo for QFX5000 series l3_switches
In recent JunOS releases it looks like the commands changed slightly
for controlling what packet headers are used to perform hashing as
part of ECMP routing.

This change ensures only src + dst IP are used for ECMP on our L3
devices, ensuring traffic between a given set of endpoints will
always be sent out the same interface.  We have quite a lot of
variance in the IPs across our network (and from outside), so
this hopefully won't lead to any major imbalance in uplink traffic
levels on these switches.  We need to probably introduce on slowly
and monitor things just to be sure.

Leaving it off for EVPN devices for now (see lats comment on
task).

Bug: T339852
Change-Id: Ie7b333757e74ba4f545ff5563d38d5e8dda66cb2