Explicitly set hash mode type on QFX5100 devices for ECMP
I hit an issue rolling out the previous patch as the QFX5100 devices
were not defaulting to "l2-payload" as the hash mode.

On the 5120 this command is not present, hashing on IP header seems
to be the only option there.  But on the two older cloudsw in C8 and
D5 we need to add this other command before the remaining config
can be applied.

Bug: T339852
Change-Id: I7f3a91efa419f603850c63b8958f633744aa784b