Add ns0 and ns1 /32 routes to anycast_prefixes list
It could be argued that combining these host routes into this list
doesn't quite make sense.  From a config point of view it will work
ok, allowing them to be announced by dns hosts.

The list is also used to add community when processing bgp_out
prefixes, but they won't be in that.  Can't see anywhere else
it will cause a problem config-wise.

Bug: T348041
Change-Id: I0b8d111d4058cfe7069efab29470edfe505afb8b