Streamline BGP neighbor definition in YAML and inclusion in templates
This patch further refines how we define BGP neighbors, both for L3
top-of-rack switches and our core routers.

The main idea is to rename the key we use in the YAML to match the
template used for the associated BGP group.  We can then loop over
the YAML defs, and include all required BGP groups and neighbors.

This should make it easier to include additional groups in future
if needed, for instance I've prepped in the JSON schema to have
definitions for management router to l3_switch peerings.

At the 'site' level all the groupings are moved under a new key,
"site_bgp", so we can iterate over all the sub-elements.

No-op on all current devices:
https://phabricator.wikimedia.org/P52916

NOTE: Also change on tox.ini to set jsonschema vserion,
see https://phabricator.wikimedia.org/T348764

Bug: T312635
Change-Id: I0bc2f7f9e515bd1e90708e805d6498c4268d4a08