Unify BGP policy on L3 and EVPN switches and adjust LVS backup pref
In the end it was best to combine these changes due to the inter-
related nature of them.  To go live some manual changes will be
required as BGP group names change on the switches, which will
cause sessions to bounce.

Main focus is on the policies for L3 and EVPN switches, bringing
the policy-options configuration on the non-EVPN switches under
automation for the first time.  The change also adjusts the BGP
sessions from switches to CRs, simplyifying the filters on the
basis we filter ingress already from servers.

In terms of LVS backup routes it adjusts our policy to match on
the MED inbound, and then sets a community and the local-pref
based on that.  The community is also matched on the EBGP
boundary between CR and Switch (both sides) to lower local
pref.  This ensures the backup routes are always treated as
such, regardless of AS-path length or specific ASNs in the
path.

Plan would be to reconfigure manually until it's a no-op, then
merge.

Bug: T344601
Bug: T349125
Bug: T348446
Change-Id: Ia3015f87577a4691a8d873aeb614286c03000eac