Change core router config to export internal routes to Switches
This new policy is aimed at exporting all internal routes from our
core routers to our L3 switches (standard and evpn).

Basic approach is to announce the following:

- OSPF routes (all 'internal' routes, i.e. no as-path)
- All BGP learnt from private ASN (i.e. other swithces, LVS, K8s etc)
- Default route (v4 & v6)

On all but the synthetic default route we set the MED attribue
to the OSPF cost (of the route itself if its coming from OSPF,
or of the next-hop if route is from BGP with next-hop known
in OSPF).  This means switches pick uplink to the 'correct'
CR router, i.e. packets won't arrive on CR1 that it then sends
on to CR2, switch will send directly to CR1.

Tested in esams, some output here:
https://phabricator.wikimedia.org/P53058

Happy to discuss alternate ways to achieve the same, or indeed the
logic of what we wish to send.  I think the approach of "sending
all internal routes" is simpler in terms of policy than trying
to pick and choose which we send, esp. since we need most of
them.  It's just under 1,000 routes in each address fam so should
be fine in terms of switch TCAM.

Bug: T344547
Change-Id: Ib8486e805aa4a3fa0c76fff388f3d0171f7eb14a