Deny traffic from cloud pub ranges to WMF private IPs and tidy conf
Now that the last cloudvirts have been moved to the dedicated cloud
racks and connect to cloud-private, they no longer need to reach
services exposed on public IPs within the cloud realm from thier
10.x private addresses (instead they use 172.20.x on the cloud-
private vlan to do so).

So moving the allow for cloud-public ranges to the end of this
policy.  It may be a bit specific, we could potentially revert
to the 'allow all' at the end, and remove the stuff blocking
all but icmp to cloud-transport.  But for now want to move
slowly.

Bug: T347030t
Change-Id: I79fcfa94dc7b5c94c52a55d48ec9f06cff3832f0