Move mr1-codfw OSPF interface to et-1/1/5 on CRs after migration
After we move the uplinks from row A switches to the new Spines
in codfw vlans 401 and 402, trunking to mr1-codfw, will be
connecting to the CRs over interface et-1/0/0 (from Spiines).

This changes the OSPF config to enable for those interfaces rather
than the old ae1 ints connected directly to asw-a-codfw.

Bug: T347191
Change-Id: I5fb6ea1b0986d8d5e56d0f82b8fdd83908c38c41