Remove temp filter stopping codfw ssw's announcing row a/b networks
Had this in place while we were doing some tests on the spine
switches in codfw.  Reason was to allow testing the config for
the row A-B vlans / irb ints when the ssw's were connected to CRs
(and announcing routes in BGP), but not yet to the ASWs (and thus
not able to reach any live devices on those vlans).

Now the Spines are bridged into the row A and B ASW virtual-chassis,
so they can reach devices on those existing vlans if needed.  It's
thus safe to remove this filter.

Bug: T347191
Change-Id: Icf9f868c8c25cc2a2e2ffa29189aa988626f06a2