Adjust BGP_Customer_out policy to send default and local POP routes
Modifies the BGP_Customer_out policy to send the prefixes we normally
send from that site (using bgp-out prefix list, similar to on
Transit and Peering), and also send a default route.

Matches config in place on cr3-ulsfo that was added to fix issue for
WMF SF Office.  Longer term to be discussed how we operate over that
link or if we should even keep it.

Bug: T350740
Change-Id: Ia5a4bdd92e9723fdef7bea2ee58d0e2f71c9e400