Add BGP to the contributing protocols for aggregate routes on CRs
Allows routes from BGP to contribute to aggregate creation.  I think
the "no_remote_confed" probably isn't needed, as no routes from a
remote confed should match the prefixes configured in the
aggregate config at any site.  But perhaps it's safer to include.

Not essential, we have loopbacks and other contributing link
networks, but seems to me public vlans and LVS should be
contributing to this.

Bug: T351456
Change-Id: Id9293a44d4dc36b0c3ae7178c002a5e3d36d94ad