Change router advertisement template to set description correctly
This template has an 'if' to ensure we don't send RAs on "XLink"
interfaces.  However due to where it was in the block it was still
adding the comment for these, even though it left out the rest.

This moves the comment within the 'if' so we don't end up with
'orphan' comment lines relating to XLink vlans in the RA stanza.

Bug: T347191
Change-Id: I3d2e6c98d0c9e6131902ff48beb17dffd75f27e5