Reset spine switch BGP to CR if max prefix tripped after 30 mins
Don't expect this will happen, but recent Optus outage has made me
more sure that a reset is a good idea on these.

It's likely to happen through human error on the CR side, and
require fixing there, but I can't see that it hurts to have this
and 30 minutes shouldn't be too excessive in terms of retries.

Bug: T349116
Change-Id: I139225c24290d9699feb9d082395b91dd52f2c88