Modify K8s BGP groups to only enable multihop on CRs
The 'multihop' statements in this group are not needed on L3
switches, and indeed cause the switch to try and use its loopback
rather than irb interface to establish the peering, which fails.

So adjusting to change so that only applies on CRs.  Also moving
the 'advertise peer-as' to the block for CRs only as routing is
optimal via the hosts default gateway when we have a L3 switch,
they don't need to receive all the other K8s routes for that.

Lastly enabling multipath on all the templates to normalize the
setup between hosts.

Bug: T357619
Change-Id: If74ada9079fe737b820d76dde6eced020aa2f037