Change name of dhcp_relay var and use it to control CR IPv6 RAs also
Changes the name of the 'dhcp_relay' var, which currently we use in
sites.yaml, to 'dhcp_relay_ra', and uses it, rather than the
interface VRRP state, to enable/disable IPv6 RAs.

This makes sense as we always want DHCP relay function and IPv6 RAs
when the CR is gateway on a vlan, so one var can be used for both.

Slight change to the var as well, removing nested 'interfaces' key
as it was the only one and it made the templates less messy.

No-op on all existing devices apart from codfw CRs, where I've
removed the et-1/0/5.2017 interface for private1-a-codfw.

Bug: T358220
Change-Id: I100ae4f50f396d95ee8873ab2ea26a9d5e8f85c3