Use loopback for DHCP relay on single-ip EVPN anycast GWs
Patch changes the dhcp_relay config to explicitly tell switches to
use their overlay loopback IP to relay DHCP requests from clients
to the install server, if those clients are connected to vlans with
a single-IP anycast GW (i.e. public ones where we don't want to
waste IPs on every switch).

For this configuration to work we also need to have the IRB GW's
actual IP added to the Option 82 block, as sub-option 5 as
described in RFC3257.  This is a global command for all relayed
DHCP, but the extra field isn't causing problems on regular vlans
so it's fine.

Bug: T358488
Change-Id: I5172fdd9ae88da1dd74bfd73f1d7faabe16fc8b2