Set AS number for BGP EVPN devices globally at site level
The proposed IBGP design as discussed on task involves only a
single common ASN used on all EVPN devices at a given site.

Given that's the case listing the AS under every switch is a
little verbose in the YAML.  This patch instead moves it to
the site level, under the existing evpn_bgp key.

Using it in the templates is a little more complex than one
would hope - a variable 'set' at the top level is not
readable from a template we include with our "section" macro,
so we need to set the asn var in each sub-template instead.

The change isn't needed, we can leave on per-device if that's
perferred, but I think it's an improvement.

Bug: T365169
Change-Id: I2acf651c79d700f7987b31b0ee31c7ce9f421d45