Change EVPN BGP YAML to group into clusters and add codfw switches
The main change here is to adjust the YAML structure for the EVPN
IBGP setup so that the devices are gouped into different sets of
RRs and Clients.  The EVPN BGP template is also updated to build
the correct set of peers based on the new structure.

This implements the IBGP setup described on task.  No-op on current
devices.

Also includes the new codfw row C and D devices so if ok I won't
merge until the cabling is complete for them.

Bug: T365169
Bug: T364095
Change-Id: Iaf402b9bb89370c9fdbea2339e2a6b0ab13521aa