config/common: update list of ntp_servers to use anycast NTP servers
Since we have moved to the anycast NTP servers, we no longer need to
maintain this list in Homer. This commit updates it to specify
ntp-[abc].anycast.wmnet and alleviates the need for future upgrades
during DNS box refreshes.

Also update policies/cr-labs.yaml in this commit as well so that we
don't forget to do it later.

To be merged after the commit that it depends on mentioned below.

Bug: T366360
Depends-On: Iec68350e16c6e2c16f0e4c14576e7d0c6756d6b9
Change-Id: I17f101e023cf16e32d96ece81145f2ea02af8a21