Re-mark external traffic to DSCP BE on CRs, and rename fwd classes
Adds terms to the border-in filter to rewrite the DCSP bits on
traffic arriving from the internet to DE/BE (00000).  We want
to have that in place before we apply configuration anywhere
else that will match on DSCP values.

Also renames the default JunOS forwarding-classes to what we
ultimately want.  The border-in filter explicity maps traffic
to the 'normal' class, so we need to have it there.  However
this is not really a change, as prior the system would map to
the 'best-effort' class (queue 0) by default and all we've done
is rename it.

Bug: T339850
Change-Id: Ia087104174d760326ea7d8c07823d60b89eddeed