Add class-of-service scheduler and classifiers plus var to control
This patch adds the required classifiers and schedulers for the QoS
design to switches and routers.  It also adds a var "diffserv_qos",
which we can use to control whether those elements are applied to
any device interfaces or not (without that the schdulers and
classifiers are much like policies that aren't applied, i.e. they
have no effect).

The 'interfaces' section uses the data from the wmf_plugin
_get_qos_interfaces() function to add the required config for
each interface.

It also adds the class-of-service for management firewalls.  The
config there is different.  Firstly as there are some syntax
differences, but also because we aren't trusting markings from the
devices, we leave the default classification (which is putting
everything bar it's own routing protocol stuff in queue 0). We
add a "rewrite" rule on the uplinks so this is properly
formatted for the classifer on switches/routers.

Example diffs:

CR:  https://phabricator.wikimedia.org/P65473
ASW: https://phabricator.wikimedia.org/P65474
LSW: https://phabricator.wikimedia.org/P65475
MR:  https://phabricator.wikimedia.org/P65472

Bug: T339850
Change-Id: Ie034be7abb39a599eb0035822a72a8eb80be4d3e