Adjust route generation for Anycast ranges at eqord
The previous patch - aimed to announce these ranges from eqdfw and
eqord, did not make a difference in eqord.

Reason being the BGP_from_anycast policy's first term blocks any
route learnt from a remote confed as contributing.

That's probably fine in eqdfw, which is in the same sub-as as codfw
and can have the same policy.  But in eqord we are in a different
sub-as, so we want different criteria for announcing the Anycast.

Seems reasonable to me to re-use the BGP_agg_net_pops policy for the
Anycast ranges in eqord, as this will allow routes from a remote
Confed to contribute, but only if they are from eqiad/codfw, which
matches I think what we want at that location.

Bug: T367439
Change-Id: Ic5e8e2fd5455051e05a3a6ba6cb44774deb7567e