Disable config for RA generation on Spines in codfw
This config is from the previous migration.  We'll actually need
to add it back shortly, to move the active GW device from CR to
the Spines.  But at this step of migration the CR should be the
only one sending.

Eventually then when all hosts are moved we can delete the IRB
ints from the Spines completely, and remove the if.

Bug: T366941
Change-Id: I6a7e6519e269e60aae9b8cc80acd0bed4538418f