Support SSW performing DHCP relay for hosts connected to ASW
During the migration from legacy ASW/VC switches to LSW we have
the ASW connected to our Spines, which have IRB interfaces
configured as anycast GW acting as gateway for the various vlans.

This means the SSW needs to perform the DHCP relay function to the
install server, however existing config for it was pushed same
as LSW, meaning it inserted it's own DHCP 82 info (ae0 port) not
the original data from the ASW.

This changes it so it's closer to the old CR setup.  Tested and
working for sretest2002.

Bug: T369274
Change-Id: I7dfcaa5eec0dfdeb0a1202c41fbfc3b6e6823116