Set DHCP relay function to 'forward-only' mode for all EVPN switches
Now that we are running JunOS 22 on all the EVPN switches in eqiad we
can enable 'forward-only' for the DHCP relay function, while still
inserting the required Option 82 info.

This is a simpler, stateless mode which is probably better overall,
and also means we don't hit the bug with ARP we previously had, plus
a few other things we hit on the public vlans in codfw.  It's been
running that way in codfw for some time with no issues.

Bug: T306421
Change-Id: I50022466f70e87fc44c0a43f82c4e05911504faf