Apply qos interface config in ulsfo and on lsw1-c6-codfw
Want to start slowly with this.  lsw1-c6-codfw is a new switch with
no servers connected, so is a good place to start with routed /
QFX5120 devices to monitor.  lsw1-d4-codfw is similar although it
has some as-yet-not-live new ganeti hosts.  It has sretest2002
which will be useful for some tests.

Also adding to the VC switch (5100s) in ulsfo and the CR routers
there.  No issues expected but will pay close attention after to
make sure everything looks healthy and then can slowly roll out
to other POPs and finally our core sites.

Bug: T339850
Change-Id: If8a2eb79d646f0cb678d27e24a1d06775ddee72f