Modify class-of-service scheduler config for qfx5100
Despite the Juniper docs stating that the QFX5100 supports 'port
scheduling' when applied the 'show configuration' output said
it was not going to apply the scheduler as the platform didn't
support it.

Path enables asw-d-codfw for interface qos config, as well as
adding the qfx5100s to the set of devices we apply the config using
the ETS-style config.

Bug: T373594
Change-Id: I891d5436ae6d2d238e686db0696161a393acfd01