Adjust cloudsw/cr bgp policies and include new IPv6 range for codfw
This changes how we build the prefix lists used in the cloudsw in/out
policies.  I've removed the 'wmcs_prefixes' definition from our
YAML files, and instead added the site-specific ranges for the
prefix lists to the individual policy templates under
templates/includes/customers, which matches how we do this
elsewhere.

That does lead to a little duplication for the cloudsw side, where
the ranges are listed again, but seemed preferable to introducing
another var in the YAML for 'wmcs_prefixes6' or similar.  Happy
to discuss that though there is no true right/wrong way here.

Bug: T245495
Change-Id: I930834bb69956cf3155fbef194e9dcc001e5720e