Modify LVS_Import policy to support both Liberica and PyBal
Current plan for testing Liberica is to use re-purposed hosts with
names starting lvs*, which will be mapped to the PyBal BGP group
by automation.

As they will set the AVOIDED_PATH (14907:0) community for backup
routes rather than the MED value, we need to adjust the policy to
match on this and downpref the routes.

Bug: T375464
Change-Id: I8986ada7439f59a90f82aa50adef0e8ac8d460c8