Remove WMCS codfw prefix from CR aggregate conf and adjust outfilter
Removes the WMCS codfw prefix from our aggregate config in codfw as
this route is now being generated by the cloudsw itself.

Also modifies the BGP6_outfilter policy so it is in line with what we
have for IPv4, only announcing customer prefixes with 1 AS-hop in
the path.  Without this the WMCS prefix would be announced from
eqiad and the pops as well.

Bug: T245495
Change-Id: I880bb2daa2e0738d8b5f683e86c5e8d477391569