Move idle-timeout under login to the dedicated login template
The pfw's include the login template as do our other devices, but
they have their own template for the overall 'system' section.

They should have this 'idle-timeout' command added to force logoff
the serial console to a password prompt.  So moving it to the
login template so they will get it as well as everything else.

Bug: T377381
Change-Id: Ibecd1f9ad7eed5917798ced5fde79d2d33734e52