Add automation for IPsec tunnels on srx devices based on Netbox
The additions to the security template adds IKE and IPsec config
elements as needed to define IPsec p2p tunnels on SRX devices.

The config is verbose as always with IPsec, but it all comes
pretty much directly from Netbox.  I've used generic naming for
everything, based on the interface name, which was the simplest
approach.  This would mean we may have some duplication (two
proposals with same ciphers defined) if we had multiple similar
tunnels, however it won't cause any issue config or performance-
wise.

Diff to existing config for codfw is here:
https://phabricator.wikimedia.org/P71000

Bug: T378020
Change-Id: I78ad38541e987a516eb0c9f48820384b451ac0c1