Update routing policy to introduce BACKUP_PATH community
These changes add a new named BGP community called 'BACKUP_PATH',
which will have a value of 14907:1.  On core BGP peerings
between our own network devices (i.e. between core routers and L3
switches) we will match this community and set the local preference
for the routes to 70.

This also adjusts the PyBal policy to set this community, rather than
AVOIDED_PATH, for routes learnt with MED=100.  So this will be the
default community used for PyBal / L4LB normal "backup" routes,
which allows traffic team to use MED=200, or set 14907:0, to
have a route given local-pref 50, which will be less preferred than
the normal 'backup' route.

Some minor renaming of the policy terms as well to be clearer.

Bug: T354839
Bug: T379164
Change-Id: I3f8f37519906a491e79a9f75d0b6a127449583ac