Class-of-service: don't insert comment with host name under cos/ints
Remove the comment we had been adding to each "interface" stanza
within the class-of-service context.

Comments can't be added with a 'set' command from what I can tell,
which means we can't do it from the switch-interfaces cookbook.
Better to remove from Homer so the produced configs won't trigger
a diff.

Also changing the conditional for our EX4600 model to just look for
'EX4'.  Not technically neccecary as Homer only connects to VC
masters which means EX4600 is the only model in that series it
interacts with, but doing for consistency with the cookbook.

Bug: T379549
Change-Id: Ic298c23b66c61cf0150c4ac1787f9db7cd17b3cf