Update policy for K8s BGP to allow a wider range of v4 prefixes
As per the task this will change the router/switch policy to
allow K8s nodes originate prefixes from a larger /17 range.

This change can be made any time as it won't affect existing
announcements.  But it MUST be done before any hosts start
announcing prefixes from the upper half of the /17.

Bug: T375845
Change-Id: Ie138f55fae6963e2a36a71c999a89d176b83e802