Cloud-in: Add specific term allowing ICMPv6 from cloud-transports
The previous patch did not do what we needed as Capirca treats
ICMP and ICMPv6 as separate protocols, so the existing top rule
to allow ICMP from the transport ranges didn't result in any
term being added to the v6 version of the filter even though
we added the IPv6 ranges to cloud-transports.

Adding new term to match ICMPv6 also so this is corrected.

Bug: T379283
Change-Id: Iddc18f98127fc336d8aa0a2257301f4b2dafdd53