Support setting custom arp-policer on CR interfaces
The MX platform has a default policer for ARP traffic to the RE,
with a reasonable but quite small limit.  As discovered in the
referenced task this was causing problems in magru, where the
level of ARP broadcasts on the IX.BR LAN is quite excessive.

Until they get this under control we need to set up a separate,
customer policer for the IX.BR interface, which will stop us
dropping ARP requests and thus stop BGP adjacencies going down
when the ARP for neighbors times out and can't be refreshed.

Bug: T384774
Change-Id: Ie7d1cd158567bd0f55a7ccbaa4f07276530973f5