prepend_as_out: switch outbound policy rather than modify existing
This patch changes our existing BGP_outfilter policy to remove the
conditional adding lines to pre-pend our AS outbound if the
prepend_as_out var is set.

Instead it adds new, static policies, BGP_outfilter_prepend and
BGP6_outfilter_prepend, which always have the lines to prepend
our AS in them.  It then moves the conditionals to the relevant
BGP groups in cr/bgp.conf, changing the export policy for those
groups if the var is set.

The benefit of this is it leaves the unmodified BGP_outfilter in
place on the device, which can easily be applied manually to
select peers if we want to stop prepending to them.

Diff: https://phabricator.wikimedia.org/P74291

Bug: T389606
Change-Id: Iff74f362ddf78312b23fe7c573011bb54055eb25