Add prepend-as-out variable for each site always
Patch to add prepend-as-out var under each site in our homer vars
always, set to false.  This is already the default, so in a way
a waste or uneeded to have it defined all the time.

However I feel in the kind of situation it might be needed, for the
SREs dealing with things, it is much more straightforward if the var
is already defined and they just need to change 'false' to 'true',
than to add it in the correct place.

Bug: T389606
Change-Id: If924a18842bfd7275892013b92ae26dca445d3eb