Cloud network: update policy to support /17 IPv4 aggregates
This patch makes a few changes to our current policies related to the
RFC1918 space used by Cloud VPS instances.  The background is the
new networks being added in eqiad outside of the legacy /21 range,
this patch introduces new aggregates better reflecting what is
allocated to cloud and permits those instead.  It also makes
several other changes:

* We redefine the 'CLOUD_VMs' def for Capirca to 172.16.0.0/16
* We separately reference the entire RFC1918 block as 172.16.0.0/12
* We remove the 'orlonger' statement in the cloud-in policy on CRs
** Cloudsw only announce the full ranges, so we can be precise
* We add the new /17 aggregates to existing prefix lists / policy

Removing the exisitng /21 and /24 range will come in a later patch,
idea is to merge this first, add the new statics on the cloudsw
side, check the aggregates are announced in addition to the existing
routes, then remove the more specifics from the setup.

Bug: T364725
Change-Id: I3a04aeb05248e76f8468059726270b9a324be0e1