WMCS: Add policy for IPv6 ranges assigned for server BGP announcement
This patch creates a new prefix list and policy intended to be used
on BGP peerings to certain cloud servers, such as cloudlb, where they
announce /128 host routes for various public and private services
offered in the 'cloud realm'.

A public and "private" (blocked on our CRs ingress from internet)
range is allocated for each core site.  To compare to Eqiad the
two ranges would be equivalent to 185.15.56.160/28 and
172.20.255.0/24.

Bug: T379282
Change-Id: I94f7acb985144ec56a55f85dd8b72e4d342aba1a