Adjust OSPF metric on cr3-ulsfo -> cr4-ulsfo 100G link
The fibres and optics on this link were replaced yesterday, and thus
far we have not had errors on the link (as we had previous), nor
has it been regularly flapping.

As such removing the higher metric on this connection to put traffic
on the link and see does it stay clean.

Keeping the separate L3 link for now if all is ok next week we can
decide how to proceed.

Bug: T390731
Change-Id: Ic5862ee84b493ab19392f03891d48befc8267c86