New device additions for codfw expansion plus policy changes
This patch adds the new switches for the codfw expansion to
devices.yaml, as well as defining the BGP peerings they need to the
spine switches in rows A-D.  Additionally it updates the BGP
config for the row A-D spines to use the new /127 IPs for EBGP.

The policy changes are outlined below:

* ibgp_out is the current policy being used on the IBGP peerings
between switches in the new cage.  It exports only routes from
IRB interfaces and those learnt from hosts in EBGP.  Link networks
and loopbacks are already in OSPF.

* core_out policy is modified as follows:
** we export /32 and /128 host routes (loopbacks and service vips)
** we export networks larger than /29 and /64 (vlan subnets)

Tested on asw1-bw27 on just the peering to CR1 (see P76368), it does
cease announcing the /31s for instance, however we don't need them,
they are either between CR and asw anyway (thus known from direct)
or it's the linknet used between ASW and MR which is not required
in the CR table (the loopback and MGMT range announced by MR are
correctly propagated).

The IBGP group on the new switches remains manual, that will require
a new structure for defining IBGP and finding neighbors.  But with
this patch in place we can use Homer against these switches to
configure ports for servers or any other elements.

Bug: T394021
Change-Id: Ib9eb4155ed55613e4a7579412171e9f4d72e8555