Switch IBGP Policy: set next-hop self on routes being originated
I'd left out the 'next-hop-self' statement for the routes we send
in IBGP, which is needed as remote devices may not have a route to
the original next-hop of an EBGP one.

This patch also puts the two policies used in the unicast IBGP
case in an 'else' to the existing if statement which adds the
EVPN-specific policy.  Result is the unicast policies are not added
to EVPN switches where they are not used.

Bug: T394530
Change-Id: If53b1ca1f8c38cda72229d03f4f6d6c95bf9c217