Templates: replace 'section' macro with include statements
Removes our macros for login and section - which wrap include
statements - and instead uses include statements directly in
the templates instead.

This allows for variables set in Jinja to pass from one template
file to another, which doesn't happen using the macro.

Quite a few of the template files I've added a 'filter indent'
statement to so we get the same indent as before.

Homer runs fine against all our devices with this and reports
no diff: https://phabricator.wikimedia.org/P77039

Bug: T395555
Change-Id: I445d00fa06599a6779dbd95e6dfc31bffd0ac04b