ASW Templates: modify Jinja templates step 1 (try 2)
This patch applies some changes to our templates in line with
other forthcoming adjustments to support IBGP.  The resulting
config is not modified on any existing devices, changes are
only in templates.

Specifically the changes are:

- Moves some evpn specific config out of 'evpn_bgp' and into a
new template file - 'evpn.conf'
- Deletes the existing switch_options.conf file and moves the
lines from it to 'evpn.conf'
- Removes the support for 32-bit ASNs when using EVPN (we don't
have this anywhere)
- Renames the 'unicast_bgp' template to 'ebgp'
- Sets the 'asn' var in top-level asw.conf and removes lines doing
it in other templates.

Bug: T394530
Change-Id: Ieb87147650d63c6fc4cb0c7ba4f8fb09f3632637